Hooks moved out to handler_utils.py
