+++ celery_serverless/handler.py
@@ -6,10 +6,10 @@ except ImportError:
 
 import os
 import sys
-import importlib
 import logging
 
-logging.basicConfig()
+from .handler_utils import maybe_call_hook, import_callable, ENVVAR_NAMES as ENVVARS
+
 logger = logging.getLogger(__name__)
 logger.propagate = True
 if os.environ.get('CELERY_SERVERLESS_LOGLEVEL'):
@@ -17,32 +17,8 @@ if os.environ.get('CELERY_SERVERLESS_LOGLEVEL'):
 print('Celery serverless loglevel:', logger.getEffectiveLevel())
 
 
-def _maybe_call_hook(envname, locals_={}):
-    func_path = os.environ.get(envname)
-    logger.debug("Trying the hook %s: '%s'", envname, func_path or '(not set)')
-    func = _import_callable(func_path)
-    return func(locals_=locals_) if func else None
-
-
-def _import_callable(name):
-    result = None
-    if name:
-        logging.info("Importing hook '%s'", name)
-        module_name, split, callable_name = name.rpartition(':')
-        module = importlib.import_module(module_name)
-        result = getattr(module, callable_name)
-    return result if callable(result) else None
-
-
-_pre_warmup_envvar = 'CELERY_SERVERLESS_PRE_WARMUP'
-_pre_handler_definition_envvar = 'CELERY_SERVERLESS_PRE_HANDLER_DEFINITION'
-_post_handler_definition_envvar = 'CELERY_SERVERLESS_POST_HANDLER_DEFINITION'
-_pre_handler_call_envvar = 'CELERY_SERVERLESS_PRE_HANDLER_CALL'
-_error_handler_call_envvar = 'CELERY_SERVERLESS_ERROR_HANDLER_CALL'
-_post_handler_call_envvar = 'CELERY_SERVERLESS_POST_HANDLER_CALL'
-
 ### 1st hook call
-_maybe_call_hook(_pre_warmup_envvar, locals())
+maybe_call_hook(ENVVARS['pre_warmup_envvar'], locals())
 
 # Get and activate some extras
 from celery_serverless.extras import discover_extras
@@ -55,7 +31,7 @@ from celery_serverless.worker_management import spawn_worker, attach_hooks
 hooks = []
 
 ### 2nd hook call
-_maybe_call_hook(_pre_handler_definition_envvar, locals())
+maybe_call_hook(ENVVARS['pre_handler_definition', locals())
 
 
 def worker(event, context):
@@ -72,7 +48,7 @@ def worker(event, context):
 
     try:
         ### 4th hook call
-        _maybe_call_hook(_pre_handler_call_envvar, locals())
+        maybe_call_hook(ENVVARS['pre_handler_call'], locals())
 
         try:
             request_id = context.aws_request_id
@@ -115,12 +91,12 @@ def worker(event, context):
             available_extras['sentry'].captureException()
 
         ### Err hook call
-        _maybe_call_hook(_error_handler_call_envvar, locals())
+        maybe_call_hook(ENVVARS['error_handler_call'], locals())
         raise
     finally:
         logger.info('END: Handle request ID: %s', request_id)
         ### 5th hook call
-        _maybe_call_hook(_post_handler_call_envvar, locals())
+        maybe_call_hook(ENVVARS['post_handler_call'], locals())
 
         if 'wdb' in available_extras:
             available_extras['wdb']['stop_trace']()
@@ -131,4 +107,4 @@ if 'sentry' in available_extras:
     worker = available_extras['sentry'].capture_exceptions(worker)
 
 ### 3rd hook call
-_maybe_call_hook(_post_handler_definition_envvar, locals())
+maybe_call_hook(ENVVARS['post_handler_definition'], locals())
